+++ AVACLI/commands/migrate.js
@@ -11,9 +11,9 @@ function migrate() {
   const migrator = new ACMigrator();
   const seeder = new ACSeeder();
   const choices = [
-    "\x1b[32m\x1b[1mSAFE\x1b[0m \x1b[3m(Only migrates zones/tables or records that don't exist yet)\x1b[0m",
-    "\x1b[33m\x1b[1mOVERWRITE\x1b[0m \x1b[3m(Migrates over your existing zones/tables and records)\x1b[0m",
-    "\x1b[31m\x1b[1mWIPE\x1b[0m \x1b[3m(Wipes your storage/database and then migrates)\x1b[0m"
+    "\x1b[32m\x1b[1mSAFE\x1b[0m       \x1b[3m(Only migrates zones/tables or records that don't exist yet)\x1b[0m",
+    "\x1b[33m\x1b[1mOVERWRITE\x1b[0m  \x1b[3m(Migrates over your existing zones/tables and records)\x1b[0m",
+    "\x1b[31m\x1b[1mWIPE\x1b[0m       \x1b[3m(Wipes your storage/database and then migrates)\x1b[0m"
   ];
   const questions = [
     {
@@ -21,16 +21,16 @@ function migrate() {
       name: "mode",
       message: "Choose a migration mode:",
       default: 0,
-      prefix: `${terminalPrefix()}\x1b[3m`,
+      prefix: `${terminalPrefix()}\x1b[34m`,
       suffix: "\x1b[0m",
       choices: choices
     },
     {
       type: "confirm",
-      name: "seed",
+      name: "populate",
       message: "Also populate?",
       default: true,
-      prefix: `${terminalPrefix()}\x1b[3m`,
+      prefix: `${terminalPrefix()}\x1b[34m`,
       suffix: "\x1b[0m"
     }
   ];
@@ -45,7 +45,7 @@ function migrate() {
       const option = options[answers.mode];
       migrator.migrate(mode[option], (success) => {
         if (success) {
-          if(answers.seed) {
+          if(answers.populate) {
             seeder.seed(mode[option], () => {
               process.exit(0);
             });