+++ AVAFoundation/AFLocalisation.js
@@ -8,22 +8,21 @@ class AFLocalisation {
 
   /**
    * 
-   * @param {String} language For example: en
-   * @param {String} region For example: UK
+   * @param {String} locale For example: en_GB
    */
-  constructor(language, region) {
+  constructor(locale) {
     var normalizedPath = `${projectPWD}/app/localisations`;
     var localisations = [];
     if (fs.existsSync(normalizedPath)) {
-      fs.readdirSync(normalizedPath).forEach(function (file) {
+      fs.readdirSync(normalizedPath).forEach((file) => {
         const extensions = file.split(".");
         if (extensions.length === 2)
           if (extensions[extensions.length - 1].toUpperCase() === "JSON")
             localisations[extensions[0]] = require(`${projectPWD}/app/localisations/${file}`);
       });
-      this.list = localisations[language];
+      this.list = localisations[locale] || {};
     } else {
-      this.list = null;
+      this.list = {};
     }
   }
 
@@ -31,21 +30,13 @@ class AFLocalisation {
   getList() {
     return this.list;
   }
+  
 
-
-  _(context) {
-    console.log(context, "=>", context);
-    return typeof this.list[context] === "string" ? this.list[context] : context;
-  };
+  translate(context) {
+    return typeof (this.list || {})[context] === "string" ? this.list[context] : context;
+  }
 
 }
-
-
-export function translate(context) {
-  const localisation = new AFLocalisation("en");
-  const locale = localisation.getList();
-  return typeof locale[context] === "string" ? locale[context] : context;
-};
 
 
 export default AFLocalisation;